requirements.txt
aioredis==2.0.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio-redis==0.16.0
certifi==2022.9.24
Deprecated==1.2.13
Django==4.1.2
django-elasticsearch-dsl==7.2.2
djangorestframework==3.14.0
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
packaging==21.3
psycopg2==2.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
